facture/l/requirements.txt
# Core dependencies
eel==0.18.2
pywebview==5.3.2
Pillow==12.1.1

# Windows build (installed automatically on Windows)
# pyinstaller==6.19.0
